:sparkles: feat(language): Add Language enum
Added the `Language` enum to handle supported languages in the project.

:bug: fix(url): Update url_launcher version

Updated the `url_launcher` package version to `6.3.0` to resolve compatibility issues.

:arrow_up: chore(deps): Update dependencies

Updated various dependencies to their latest versions for improved performance and compatibility.

:recycle: refactor(main): Refactor main.dart

Refactored the `main.dart` file to improve code readability and organization.

:sparkles: feat(localization): Add getLocalizationDelegate function

Added the `getLocalizationDelegate` function to dynamically retrieve the localization delegate based on the device's language.

:art: style(README): Update README formatting

Updated the formatting and content of the README file for better clarity and consistency.

Signed-off-by: Dmytro Turskyi <dmytro.turskyi@gmail.com>